chore: finalize helm resource repo refactor
On-behalf-of: SAP <matthias.bruns@sap.com>
