+++ bindings/go/helm/blob/chart_blob.go
@@ -87,7 +87,10 @@ func extractFromTar(tarBlob blob.ReadOnlyBlob) (chartBlob blob.ReadOnlyBlob, pro
 		}
 	}()
 
-	tr := tar.NewReader(rc)
+	// make sure to get all from pipe (won't work :D )
+	tarBytes, err := io.ReadAll(rc)
+
+	tr := tar.NewReader(bytes.NewReader(tarBytes))
 	for {
 		hdr, err := tr.Next()
 		if err != nil {
