+++ bindings/go/helm/cmd/main.go
@@ -14,6 +14,7 @@ import (
 	genericv1 "ocm.software/open-component-model/bindings/go/configuration/generic/v1/spec"
 	constructorruntime "ocm.software/open-component-model/bindings/go/constructor/runtime"
 	helminput "ocm.software/open-component-model/bindings/go/helm/input"
+	inputspec "ocm.software/open-component-model/bindings/go/helm/spec/input"
 	helmv1 "ocm.software/open-component-model/bindings/go/helm/spec/input/v1"
 	plugin "ocm.software/open-component-model/bindings/go/plugin/client/sdk"
 	"ocm.software/open-component-model/bindings/go/plugin/manager/contracts"
@@ -76,7 +77,7 @@ func main() {
 		Level: slog.LevelDebug,
 	}))
 
-	capabilities := endpoints.NewEndpoints(helminput.Scheme)
+	capabilities := endpoints.NewEndpoints(inputspec.Scheme)
 	helmPlugin := &HelmInputPlugin{}
 	if err := input.RegisterInputProcessor(&helmv1.Helm{}, helmPlugin, capabilities); err != nil {
 		logger.Error("failed to register helm input plugin", "error", err.Error())
