+++ bindings/go/helm/spec/input/scheme.go
@@ -1,0 +1,15 @@
+package input
+
+import (
+	v1 "ocm.software/open-component-model/bindings/go/helm/spec/input/v1"
+	"ocm.software/open-component-model/bindings/go/runtime"
+)
+
+var Scheme = runtime.NewScheme()
+
+func init() {
+	Scheme.MustRegisterWithAlias(&v1.Helm{},
+		runtime.NewVersionedType(v1.Type, v1.Version),
+		runtime.NewUnversionedType(v1.Type),
+	)
+}
